requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
flake8==3.9.2
httpie==2.4.0
idna==3.2
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==2.1.0
PySocks==1.7.1
pytz==2021.1
requests==2.26.0
requests-toolbelt==0.9.1
sqlparse==0.4.1
urllib3==1.26.6
